+++ apps/api/src/mail/emails/signin-code-email.tsx
@@ -1,0 +1,42 @@
+import { Section, Text } from '@react-email/components'
+import { otpStyle, text, workspaceDetails } from './styles/common-styles'
+import BaseEmailTemplate from './components/base-email-template'
+
+interface SignInCodeEmailTemplateProps {
+  code: string
+  name?: string
+  device: string
+  location: string
+}
+
+export const SignInCodeEmailTemplate = ({
+  code,
+  name,
+  device,
+  location
+}: SignInCodeEmailTemplateProps) => {
+  return (
+    <BaseEmailTemplate
+      previewText="Your Sign in Code for Keyshade CLI"
+      heading="Keyshade Sign in Code"
+    >
+      <Text style={text}>Hey {name ?? 'there'}!,</Text>
+      <Text style={text}>
+        We’ve sent you this email to verify your Keyshade account from{' '}
+        <strong>{location}</strong> using <strong>{device}</strong>. Your Sign
+        in Code is:
+      </Text>
+      <Section style={workspaceDetails}>
+        <Text style={otpStyle}>
+          <strong>{code ?? '0000'}</strong>
+        </Text>
+      </Section>
+      <Text style={text}>
+        This Sign in Code will expire in <strong>5 minutes</strong>. Please use
+        it to complete your action on Keyshade CLI.
+      </Text>
+    </BaseEmailTemplate>
+  )
+}
+
+export default SignInCodeEmailTemplate