+++ apps/api/src/mail/services/mock.service.ts
@@ -29,6 +29,18 @@ export class MockMailService implements IMailService {
     )
   }
 
+  async sendSignInCode(
+    email: string,
+    code: string,
+    name: string,
+    device: string,
+    location: string
+  ): Promise<void> {
+    this.log.log(
+      `Sign-in code for ${email} is ${code} and name is ${name}. Device is ${device} and location is ${location}`
+    )
+  }
+
   async adminUserCreateEmail(email: string): Promise<void> {
     this.log.log(`Create pAdmin User Email: ${email}`)
   }