+++ apps/api/src/prisma/migrations/20251006124352_add_unique_constraint_to_pat/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - A unique constraint covering the columns `[name,userId]` on the table `PersonalAccessToken` will be added. If there are existing duplicate values, this will fail.
+
+*/
+-- CreateIndex
+CREATE UNIQUE INDEX "PersonalAccessToken_name_userId_key" ON "PersonalAccessToken"("name", "userId");