+++ apps/platform/src/components/billing/currentBillDetails/index.tsx
@@ -4,8 +4,8 @@ import RenewSubscription from '../renewSubscription'
 import BillingDetail from './billingDetail'
 import DownloadInvoice from './downloadInvoice'
 import CancelPlan from './canclePlan'
-import { selectedWorkspaceAtom } from '@/store'
 import Visible from '@/components/common/visible'
+import { selectedWorkspaceAtom } from '@/store'
 
 export default function CurrentBillDetails() {
   const currentWorkspace = useAtomValue(selectedWorkspaceAtom)